# src/resources/users/schemas/user.schema.ts
import { Exclude, Transform } from 'class-transformer';
export class User {
  id: string;
  login: string;

  @Exclude()
  password: string;

  version?: number;

  @Transform(({ value }) => +new Date(value))
  createdAt?: Date;

  @Transform(({ value }) => +value)
  updatedAt?: Date;

  constructor(partial: Partial<User>) {
    Object.assign(this, partial);
  }
}